Fully migrate all commands
