+++ src/bot_extension.py
@@ -10,18 +10,19 @@ load_dotenv()
 refresh_rate = float(environ.get("REFRESH_RATE"))
 
 
-class BotExtension:
+class BotExtension(commands.Cog): # TODO: turn to cog
     '''Tasks and listeners'''
 
     def __init__(self, bot: commands.Bot) -> None:
         self.bot = bot
-
+    
     @tasks.loop(minutes=refresh_rate)
     async def refresh_role_loop(self):
         '''Refresh all roles, periodically'''
         await guild_funcs.refresh_roles_of_bot(bot=self.bot)
         print("Refreshed all guilds on: ", datetime.now())
 
+    @commands.slash_command(name = "help")
     async def help_cmd(self, inter: disnake.CommandInteraction):
         '''/help: Show this help message'''
         msg = 'Here are several things I can do:'
@@ -37,28 +38,25 @@ class BotExtension:
 
         await inter.response.send_message(msg + "```" + "\n".join(help_msg) + "```")
 
+    @commands.Cog.listener()
     async def on_ready(self):
         '''Notify the user that the bot has logged in and start to periodically refresh roles'''
         print("Logged in")
         if not self.refresh_role_loop.is_running():  # pylint: disable=no-member
             self.refresh_role_loop.start()  # pylint: disable=no-member
+    
+    @commands.Cog.listener()
+    async def on_guild_join(guild: disnake.Guild):
+        '''Add the bot to a guild'''
+        await guild_funcs.create_roles_in_guild(guild)
 
+    @commands.Cog.listener()
+    async def on_guild_remove(guild: disnake.Guild):
+        '''Remove the bot from a guild'''
+        guild_funcs.remove_guild_data(guild.id)
 
-async def on_guild_join(guild: disnake.Guild):
-    '''Add the bot to a guild'''
-    await guild_funcs.create_roles_in_guild(guild)
-
-
-async def on_guild_remove(guild: disnake.Guild):
-    '''Remove the bot from a guild'''
-    guild_funcs.remove_guild_data(guild.id)
 
 
 def setup(bot: commands.Bot):
     '''Add bot listeners and help cmd'''
-    bot.event(on_guild_join)
-    bot.event(on_guild_remove)
-
-    instance = BotExtension(bot)
-    bot.event(instance.on_ready)
-    bot.slash_command(name="help")(instance.help_cmd)
+    bot.add_cog(BotExtension(bot))
